Created SaleBoardScreen and ScrollPane
