+++ src/main/java/screen/ViewScreen.java
@@ -3,6 +3,6 @@ import javax.swing.*;
 
 public class ViewScreen {
     public static void main(String[] args) {
-        new LoginScreen();
+        new SaleBoardScreen();
     }
 }
